% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelmixer.R
\name{excelmixer}
\alias{excelmixer}
\title{excelmixer}
\usage{
excelmixer(df, file, sheet_name = "Sheet1", merge_labels = TRUE)
}
\arguments{
\item{df}{Data frame to export.}

\item{file}{Output Excel file path.}

\item{sheet_name}{Name of the Excel sheet. Default is "Sheet1".}

\item{merge_labels}{Logical, should demo cells in the first row be merged? Default TRUE.}
}
\value{
None; writes Excel file.
}
\description{
Export a data frame to Excel with mixed character and numeric values
}
