% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelmixer.R
\name{fix_guts}
\alias{fix_guts}
\title{fix_guts
Internal helper function to process data frame so numbers remain
as numbers and text as text when exporting. Borrowed and adapted from:
\url{https://github.com/jjmoncus/wink/blob/main/R/write_banners.R}}
\usage{
fix_guts(data)
}
\arguments{
\item{data}{A data frame or tibble to process.}
}
\value{
A list of row vectors, with numeric values as numeric and text as character.
}
\description{
fix_guts
Internal helper function to process data frame so numbers remain
as numbers and text as text when exporting. Borrowed and adapted from:
\url{https://github.com/jjmoncus/wink/blob/main/R/write_banners.R}
}
\keyword{internal}
